refactor(tests): P2-F04 split test_rules_structure.py into 14 focused files
Extract 13 test classes and 14 standalone functions from the 1,620-line
monolith into per-rule files. Delete the original. All tests preserved;
integration predicate class already has the P3-F14 class-scoped fixture.
